+++ pyout/elements.py
@@ -38,26 +38,45 @@ schema = {
                       {"$ref": "#/definitions/interval"}],
             "default": False,
             "scope": "field"},
+        "width_type": {
+            "description": "Type for numeric values in 'width'",
+            "oneOf": [{"type": "integer", "minimum": 1},
+                      {"type": "number",
+                       "exclusiveMinimum": 0,
+                       "exclusiveMaximum": 1}]},
         "width": {
             "description": """Width of field.  With the default value, 'auto',
             the column width is automatically adjusted to fit the content and
             may be truncated to ensure that the entire row fits within the
             available output width.  An integer value forces all fields in a
-            column to have a width of the specified value. In addition, an
-            object can be specified.  Its 'min' and 'max' keys specify the
-            minimum and maximum widths allowed, whereas the 'width' key
-            specifies a fixed width.  The 'marker' key specifies the marker
-            used for truncation ('...' by default).  Where the field is
-            truncated can be configured with 'truncate': 'right' (default),
-            'left', or 'center'.""",
-            "oneOf": [{"type": "integer"},
+            column to have a width of the specified value.
+
+            In addition, an object can be specified.  Its 'min' and 'max' keys
+            specify the minimum and maximum widths allowed, whereas the 'width'
+            key specifies a fixed width.  The values can be given as an integer
+            (representing the number of characters) or as a fraction, which
+            indicates the proportion of the total table width (typically the
+            width of your terminal).
+
+            The 'marker' key specifies the marker used for truncation ('...' by
+            default).  Where the field is truncated can be configured with
+            'truncate': 'right' (default), 'left', or 'center'.
+
+            The object can also include a 'weight' key.  Conceptually,
+            assigning widths to each column can be (roughly) viewed as each
+            column claiming _one_ character of available width at a time until
+            a column is at its maximum width or there is no available width
+            left.  Setting a column's weight to an integer N makes it claim N
+            characters each iteration.""",
+            "oneOf": [{"$ref": "#/definitions/width_type"},
                       {"type": "string",
                        "enum": ["auto"]},
                       {"type": "object",
                        "properties": {
-                           "max": {"type": ["integer", "null"]},
-                           "min": {"type": ["integer", "null"]},
-                           "width": {"type": "integer"},
+                           "max": {"$ref": "#/definitions/width_type"},
+                           "min": {"$ref": "#/definitions/width_type"},
+                           "width": {"$ref": "#/definitions/width_type"},
+                           "weight": {"type": "integer", "minimum": 1},
                            "marker": {"type": ["string", "boolean"]},
                            "truncate": {"type": "string",
                                         "enum": ["left",