+++ pyout/tests/test_tabular.py
@@ -762,6 +762,41 @@ def test_tabular_write_autowidth_min_max_with_header():
     assert_contains_nc(lines1, "bar  BAD!!...")
 
 
+def test_tabular_write_autowidth_min_frac():
+    out = Tabular(style={"width_": 12,
+                         "name": {"width": {"min": 0.5}}})
+    out(OrderedDict([("name", "foo"),
+                     ("status", "unknown")]))
+
+    # 0.5 of table width => 6 characters for "foo"
+    assert out.stdout == "foo    un...\n"
+
+
+def test_tabular_write_autowidth_max_frac():
+    out = Tabular(style={"width_": 12,
+                         "name": {"width": {"max": 0.5}}})
+    out(OrderedDict([("name", "foo"),
+                     ("status", "ok")]))
+
+    # 0.5 of table width => 6 characters for "foo", but it only needs 3.
+    assert out.stdout == "foo ok\n"
+
+    out(OrderedDict([("name", "longerthanmax"),
+                     ("status", "ko")]))
+
+    lines0 = out.stdout.splitlines()
+    # Value over 6 only takes up 6.
+    assert_contains_nc(lines0, "lon... ko")
+
+
+def test_tabular_write_fixed_width_frac():
+    out = Tabular(style={"width_": 20,
+                         "name": {"width": 0.4}})
+    out(OrderedDict([("name", "foo"),
+                     ("status", "ok")]))
+    assert out.stdout == "foo      ok\n"
+
+
 def test_tabular_write_autowidth_different_data_types_same_output():
     out_dict = Tabular(["name", "status"],
                        style={"header_": {},
@@ -798,12 +833,24 @@ def test_tabular_fixed_width_exceeds_total():
         out(OrderedDict([("name", ""), ("status", "")]))
 
 
+def test_tabular_number_of_columns_exceeds_total_width():
+    cols = ["a", "b", "c", "d"]
+    out = Tabular(columns=cols, style={"width_": 3})
+    with pytest.raises(StyleError):
+        out([c + "val" for c in cols])
+
+
 def test_tabular_auto_width_exceeds_total():
-    out = Tabular(style={"width_": 13})
-    out(OrderedDict([("name", "foobert"),
-                     ("status", "okiguess"),
-                     ("path", "illbedropped:(")]))
-    assert out.stdout == "foobert o... \n"
+    out = Tabular(style={"width_": 13,
+                         "default_": {"width": {"marker": "…"}}})
+    out(OrderedDict([("name", "abcd"),
+                     ("status", "efghi"),
+                     ("path", "jklm")]))
+    # The values are divided evenly.  Subtracting the separators, there are 11
+    # available spaces.  'status' and 'path' get 4, while 'name' gets the
+    # remaining 3.  'name' is shorted just because the columns are processed in
+    # reverse alphabetical order.
+    assert out.stdout == "ab… efg… jklm\n"
 
 
 def test_tabular_auto_width_exceeds_total_multiline():
@@ -813,12 +860,13 @@ def test_tabular_auto_width_exceeds_total_multiline():
                      ("path", "t/")]))
     assert out.stdout == "abcd efg t/\n"
 
-    # name gets truncated because it claims the most width in the table so far.
-    out(OrderedDict([("name", "notme"),
+    # name gets truncated due to predictable but arbitrary reverse alphabetical
+    # sorting when assigning widths.
+    out(OrderedDict([("name", "mooost"),
                      ("status", "metoo"),
                      ("path", "here")]))
     lines0 = out.stdout.splitlines()
-    assert_contains_nc(lines0, "n... metoo here")
+    assert_contains_nc(lines0, "m... metoo here")
 
     out(OrderedDict([("name", "hi"),
                      ("status", "jk"),